Moved spawn utils.
