+++ src/engine/core/managers/drop/DropManager.ts
@@ -8,8 +8,9 @@ import { abort } from "@/engine/core/utils/assertion";
 import { isArtefact, isGrenade, isWeapon } from "@/engine/core/utils/class_ids";
 import { parseNumbersList, parseStringsList } from "@/engine/core/utils/ini";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { getObjectCommunity, setItemCondition, spawnItemsForObject } from "@/engine/core/utils/object";
+import { getObjectCommunity, setItemCondition } from "@/engine/core/utils/object";
 import { isAmmoSection, isExcludedFromLootDropItemSection, isLootableItemSection } from "@/engine/core/utils/section";
+import { spawnItemsForObject } from "@/engine/core/utils/spawn";
 import { logicsConfig } from "@/engine/lib/configs/LogicsConfig";
 import { communities, TCommunity } from "@/engine/lib/constants/communities";
 import { TInventoryItem } from "@/engine/lib/constants/items";
