+++ src/engine/core/objects/binders/physic/PhysicObjectItemBox.ts
@@ -4,7 +4,7 @@ import { PH_BOX_GENERIC_LTX } from "@/engine/core/database";
 import { abort } from "@/engine/core/utils/assertion";
 import { parseNumbersList, parseStringsList, readIniString } from "@/engine/core/utils/ini";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { spawnItemsForObject } from "@/engine/core/utils/object/object_spawn";
+import { spawnItemsForObject } from "@/engine/core/utils/spawn";
 import { TInventoryItem } from "@/engine/lib/constants/items";
 import { ClientObject, IniFile, LuaArray, Optional, TCount, TName, TProbability, TSection } from "@/engine/lib/types";
