+++ src/engine/core/ui/debug/sections/DebugSpawnSection.ts
@@ -7,11 +7,7 @@ import { isGameStarted } from "@/engine/core/utils/game";
 import { getSimulationGroupSections, getStalkerSections } from "@/engine/core/utils/ini";
 import { LuaLogger } from "@/engine/core/utils/logging";
 import { getNearestServerObject } from "@/engine/core/utils/object";
-import {
-  getInventoryNameForItemSection,
-  spawnCreatureNearActor,
-  spawnSquadInSmart,
-} from "@/engine/core/utils/object/object_spawn";
+import { getInventoryNameForItemSection, spawnCreatureNearActor, spawnSquadInSmart } from "@/engine/core/utils/spawn";
 import { resolveXmlFile } from "@/engine/core/utils/ui";
 import { TInventoryItem } from "@/engine/lib/constants/items";
 import { LuaArray, Optional, ServerObject, TPath, TSection, Vector2D } from "@/engine/lib/types";
