+++ src/engine/core/utils/object/index.ts
@@ -10,7 +10,6 @@ export * from "@/engine/core/utils/object/object_set";
 export * from "@/engine/core/utils/object/object_setup";
 export * from "@/engine/core/utils/object/object_smart_cover";
 export * from "@/engine/core/utils/object/object_sound";
-export * from "@/engine/core/utils/object/object_spawn";
 export * from "@/engine/core/utils/object/object_state";
 export * from "@/engine/core/utils/object/object_task_reward";
 export * from "@/engine/core/utils/object/object_weapon";
