+++ src/engine/core/utils/scheme/scheme_initialization.ts
@@ -14,7 +14,6 @@ import { readIniNumber, readIniString } from "@/engine/core/utils/ini";
 import { ISmartTerrainJobDescriptor } from "@/engine/core/utils/job";
 import { LuaLogger } from "@/engine/core/utils/logging";
 import { getObjectSmartTerrain } from "@/engine/core/utils/object/object_location";
-import { spawnItemsForObject } from "@/engine/core/utils/object/object_spawn";
 import { ERelation } from "@/engine/core/utils/relation";
 import { emitSchemeEvent } from "@/engine/core/utils/scheme/scheme_event";
 import {
@@ -23,6 +22,7 @@ import {
   getSectionToActivate,
 } from "@/engine/core/utils/scheme/scheme_logic";
 import { disableObjectBaseSchemes } from "@/engine/core/utils/scheme/scheme_setup";
+import { spawnItemsForObject } from "@/engine/core/utils/spawn";
 import { logicsConfig } from "@/engine/lib/configs/LogicsConfig";
 import { TInventoryItem } from "@/engine/lib/constants/items";
 import { MAX_U16 } from "@/engine/lib/constants/memory";
